p2p: keep DA prefetch caps in relay state
- keep prefetch accounting inside DA relay state so PR scope stays Go-only

- release fulfilled reservations before byte accounting

- keep triggering peer preferred without dropping fallback peers
